Don't warn that reference list is uneditable
In articles with references, a reference list is generated. If you
try to visual edit the reference list, it displays a message "This
reference list is generated by a template, and for now can only be
edited in source mode." However it is actually editable now that
T54750 (Community Wishlist 2023 wish #2) is complete.

Bug: T54750
Change-Id: Id8115ae6045f371e4619c85aeb610fe78927d802