Add group name in data attribute to <references>'s html output
This allows using CSS to amend the references based on the
group name without wrapping the <references /> tag in another
tag (like a <div>). Essentially, you can then do `<references
group="lower-alpha" />` instead of `{{reflist|group=lower-alpha}}`
and get the same effect, which should work better in the Visual
Editor.

Using a data attribute instead of a class (as requested in
T115894) because group names can include characters that
aren't valid in CSS selectors, so we would need to amend the
group name in order to use it as a class name; using a data
attribute instead lets us keep the group name as it is.

This patch brings the legacy code back in sync with the Parsoid
implementation that already adds the same data attribute since
2014, see patch I21fbbd3, tracked via T86782. (Note Parsoid was
JavaScript back then, later cross-compiled to PHP.)

Bug: T196828
Change-Id: I91eb3ae7d441d7b87a57da5535df28de6e516bb0