Try more specific CSS selectors for clicks in browser tests
The main change here is the "enabled". Don't try to click the button
as long as it is disabled. This is critical because VisualEditor
takes a while to boot. The toolbar is already visible, but disabled.

Warning, untested! It makes sense in my head but I'm not sure if it
works.

Bug: T360026
Change-Id: Ib46d420a56effd4b4a0e48e2121106a830e5f51c