Fix preview of parent reference in context item
I believe this was a mistake in the previous patch Idd57997. Using
only the .text() of the parent reference is not sufficient. Think of
a reference that contains a <math> formula or – even worse – is
nothing but a <math> formula. The preview will be empty in this case.

Click handlers in the preview need to be disabled. It's only a
preview.

Bug: T367030
Change-Id: I7e82e03ecfeb4e9cf132985684cff5e191506307