[refactor] Don't setup the edit panel when re-using a ref
In the setup process the dialog is loaded and initialized. For the
case where the user opens this dialog to directly re-use a
reference, it does not make sense to fill out the form or create
the internal ref model. The edit panel is not used then.

This method might also be an entry point to create a new sub-ref
when triggered from e.g. Citoid. Switching to the re-use panel
still would not require to setup the edit panel.

I'm also moving a setup for the abilities, that's only relevant
in the edit case.

Bug: T375156
Change-Id: I061b88abc6cfd702a53208bac76be7a2ed6812c2